+++ sm/fingerprint.c
@@ -149,8 +149,14 @@ gpgsm_get_short_fingerprint (ksba_cert_t cert, unsigned long *r_high)
 
   gpgsm_get_fingerprint (cert, GCRY_MD_SHA1, digest, NULL);
   if (r_high)
-    *r_high = ((digest[12]<<24)|(digest[13]<<16)|(digest[14]<< 8)|digest[15]);
-  return ((digest[16]<<24)|(digest[17]<<16)|(digest[18]<< 8)|digest[19]);
+    *r_high = (((unsigned long)digest[12]<<24)
+               |(digest[13]<<16)
+               |(digest[14]<< 8)
+               |digest[15]);
+  return (((unsigned long)digest[16]<<24)
+          |(digest[17]<<16)
+          |(digest[18]<<8)
+          |digest[19]);
 }
